refactor: chatroom 기능 message 기능 분리, 중복인원으로 생성안되도록 수정
